logisitic regression and tons of plotting code for cosyne poster

+++ cosyne_abstract/plot_placecells.py
@@ -14,7 +14,7 @@ if __name__ == '__main__':
     mice = ['4139219.2','4139219.3','4139224.2','4139224.3','4139224.5',
     '4139251.1','4139260.1','4139260.2','4139261.2','4139265.3','4139265.4',
     '4139265.5','4139266.3']
-    mice = ['4139266.3']
+    # mice = ['4139266.3']
 
     df = pp.load_session_db()
     df = df[df['RewardCount']>20]
@@ -37,15 +37,16 @@ if __name__ == '__main__':
             # df_sess = df_mouse[df_mouse['Track'].str.match('TwoTower_noTimeout') | df_mouse['Track'].str.match('TwoTower_Timeout')]
             for i in range(df_mouse.shape[0]):
                 sess = df_mouse.iloc[i]
+
                 try:
                     fname = "%s\\%s_%s_%d_" % (dirbase,mouse,sess['DateFolder'],sess['SessionNumber'])
                     print(fname)
-
-                    FR, masks, SI = pc.single_session(sess,savefigs=True,
-                                        deconv=True,fbase =fname, win_trial_perm=True,
-                                        cell_method='s2p',method='bootstrap')
-                    results = {'FR':FR, 'masks':masks,'SI':SI}
-                    with open(fname+"results.pkl",'wb') as f:
-                        pickle.dump(results,f)
+                    if ~os.path.isfile(fname+'results.pkl'):
+                        FR, masks, SI = pc.single_session(sess,savefigs=True,
+                                            deconv=True,fbase =fname, win_trial_perm=True,
+                                            cell_method='s2p',method='bootstrap')
+                        results = {'FR':FR, 'masks':masks,'SI':SI}
+                        with open(fname+"results.pkl",'wb') as f:
+                            pickle.dump(results,f)
                 except:
                     print(sess)
